story(Deployment window) : file moved and test case added
